Backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.14
google-auth==2.48.0
idna==3.11
inflection==0.5.1
packaging==26.0
psycopg2-binary==2.9.10
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
